Code split & housekeeping.
