+++ src/server.cpp
@@ -5,13 +5,19 @@
  *  @author Moonsik Park, Korea Institute of Science and Technology
  **/
 
-#include <common.h>
+#include <arpa/inet.h>
 #include <fcntl.h>
 #include <netinet/in.h>
 #include <sys/socket.h>
 #include <sys/stat.h>
 #include <unistd.h>
 
+#include <cstring>
+#include <thread>
+
+#include "base/camera_manager.h"
+#include "base/exceptions/lock_timeout.h"
+#include "base/scoped_timer.h"
 #include "base/video/frame_queue.h"
 
 int socket_send_blocking(int targetfd, uint8_t *buf, size_t size) {
@@ -110,7 +116,7 @@ void socket_client_thread(int targetfd, std::shared_ptr<FrameQueue> frame_queue,
                           std::shared_ptr<CameraManager> cameramgr,
                           std::shared_ptr<types::AVCodecContextManager> ctxmgr,
                           std::atomic<bool> &shutdown_requested) {
-  set_thread_name(std::string("socket_client=") + std::to_string(targetfd));
+  // set_thread_name(std::string("socket_client=") + std::to_string(targetfd));
   int ret = 0;
   tlog::info() << "socket_client_thread (fd=" << targetfd << "): Spawned.";
   while (!shutdown_requested) {
@@ -179,7 +185,7 @@ void socket_manage_thread(std::string renderer,
                           std::shared_ptr<CameraManager> cameramgr,
                           std::shared_ptr<types::AVCodecContextManager> ctxmgr,
                           std::atomic<bool> &shutdown_requested) {
-  set_thread_name(std::string("socket_manage=") + renderer);
+  // set_thread_name(std::string("socket_manage=") + renderer);
   int error_times = 0;
   while (!shutdown_requested) {
     std::stringstream renderer_parsed(renderer);
@@ -243,7 +249,7 @@ void socket_main_thread(std::vector<std::string> renderers,
                         std::shared_ptr<CameraManager> cameramgr,
                         std::shared_ptr<types::AVCodecContextManager> ctxmgr,
                         std::atomic<bool> &shutdown_requested) {
-  set_thread_name("socket_main");
+  // set_thread_name("socket_main");
   std::vector<std::thread> threads;
 
   tlog::info() << "socket_main_thread: Connecting to renderers.";
